app/requirements.txt
fastapi
uvicorn
pandas
kafka-python
pydantic
psycopg[binary]
ipykernel
databases[postgresql]
python-dotenv
psycopg-pool

